requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
asttokens==2.4.1
beautifulsoup4==4.12.3
binwalk==2.3.4
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.37.5
botocore==1.37.5
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
crypt_r==3.13.1
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnf==4.22.0
executing==2.2.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.16.1
fonttools==4.53.1
fs==2.4.16
gbinder-python==1.1.2
httpie==3.2.3
humanize==4.9.0
idna==3.7
initial-setup==0.3.101
ipython==8.26.0
iso639==0.1.4
jedi==0.19.1
jmespath==1.0.1
kiwisolver==1.4.8
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
markdown-it-py==3.0.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mercurial==6.8.2
multidict==6.0.5
networkx==3.3
nftables==0.1
ntpsec==1.2.3
numpy==2.2.0
olefile==0.47
packaging==24.2
pandas==2.2.3
parso==0.8.4
pcapy==0.11.5.dev0
pcp==5.0
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
ply==3.11
pooch==1.8.2
productmd==1.43
prompt_toolkit==3.0.41
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pycairo==1.25.1
pyclip==0.7.0
pycparser==2.20
pycups==2.0.4
pycurl==7.45.3
pyenchant==3.2.2
pygdbmi==0.11.0.0
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyparsing==3.1.2
pyparted==3.13.0
PyQt6==6.8.1.dev2502011625
PyQt6_sip==13.9.1
PySocks==1.7.1
python-augeas==1.1.0
python-bitcoinrpc==1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2024.2
pyudev==0.24.3
PyX==0.16
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rpm==4.20.0
s3transfer==0.11.3
scapy==2.5.0
SciPy==1.11.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sentry-sdk==2.17.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.8.2
soupsieve==2.6
stack_data==0.6.3
systemd-python==235
traitlets==5.14.3
tzdata==2024.2
UPnPy==1.1.8
urllib3==1.26.20
virtualenv==20.28.0
wcwidth==0.2.13
xkbregistry==0.3
zombie-imp==0.0.2
